Moved ContextChange into a module
